updates to page 4, Experimental Optimization, to connect new Entropy object, ongoing.

+++ sans_app/support/app_functions.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import hashlib
 import os
 from PIL import Image
 import pandas
@@ -15,28 +16,72 @@ from typing import Optional, Dict, Any
 from sans_app.support import configuration
 
 
-def get_info_from_runfile(model_name, model_dir, file_dir, fit_dir):
+def process_runfile(model_name, model_dir, file_dir, fit_dir, force=True,
+                    correct_data_paths=False):
+    """
+    Parses a run script and extracts parameters and data filenames from it. Given a fit_dir, it copies the data files
+    and the script file to this directory from model_dir and file_dir, respectively. All other contents from the fit_dir
+    are deleted.
+
+    :param model_name: (str) The namem of the model, including file extension.
+    :param model_dir: (str or Path-like) The directory of the model.
+    :param file_dir: (str or Path-like) The directory of the data files.
+    :param fit_dir: (str or Path-like) The directory in which the fit problem will be initialized.
+    :param force: (bool) Whether to overwrite existing files in the directory.
+    :param correct_data_paths: (bool) Whether to correct data file paths from absolute to relative (SansApp constraint).
+    :return: (Pandas dataframe) of all parameters, (list) of all parameter names, (list) of all data filenames,
+             (bumps problem) an instance of the fit object
+    """
+
+    def _file_hash(path: Path) -> str:
+        h = hashlib.sha256()
+        with path.open("rb") as f:
+            for block in iter(lambda: f.read(1024 * 1024), b""):
+                h.update(block)
+        return h.hexdigest()
+
     file_dir = Path(file_dir).expanduser().resolve()
     model_dir = Path(model_dir).expanduser().resolve()
     runfile = model_dir / model_name
+    runfile_dest = fit_dir / model_name
 
     # extract name of data files from runfile
     datafile_names = api_sasview.extract_data_filenames_from_runfile(runfile=str(runfile))
-
     # strip any long path from filename and retain only the basename, write back to file
     datafile_names = [Path(file).name for file in datafile_names]
-    api_sasview.write_data_filenames_to_runfile(runfile=str(runfile), filelist=datafile_names)
+    datafile_paths = [str(file_dir / Path(file).name) for file in datafile_names]
+    # This modifies the original file enforcing a SANS app constraint that the data file and the script are in the same
+    # folder
+    if correct_data_paths:
+        api_sasview.write_data_filenames_to_runfile(runfile=str(runfile), filelist=datafile_names)
 
-    # check data files
-    datafile_names_user = [str(file_dir / Path(file).name) for file in datafile_names]
-    molstat.prepare_fit_directory(fitdir=str(fit_dir), runfile=str(runfile), datafile_names=datafile_names_user)
-    # if datafiles exist in user filedir, use those; otherwise create dummy files
-    for filename in datafile_names_user:
-        if Path(filename).is_file():
-            if not (fit_dir / Path(filename).name).is_file():
-                shutil.copyfile(filename, fit_dir / Path(filename).name)
-        else:
-            api_sasview.write_dummy_sans_file(str(fit_dir / Path(filename).name))
+    already_prepared = False
+    # check if model script has already been copied and is unchanged
+    if runfile_dest.is_file() and _file_hash(runfile) == _file_hash(runfile):
+        # same for each data file
+        for fname in datafile_names:
+            if (file_dir / fname).is_file():
+                if not (fit_dir / fname).is_file() and _file_hash(file_dir / fname) == _file_hash(fit_dir / fname):
+                    break
+            # no source data file is o.k., but then a dummy file of the same name should be present
+            elif not (fit_dir / fname).is_file():
+                break
+        already_prepared = True
+
+    if not already_prepared:
+        if not force:
+            if fit_dir.is_dir() and any(fit_dir.iterdir()):
+                st.info("Optimization folder is not empty. Please archive and clear contents.")
+                st.stop()
+
+        molstat.prepare_fit_directory(fitdir=str(fit_dir), runfile=str(runfile), datafile_names=datafile_paths)
+        # if datafiles exist in user filedir, use those; otherwise create dummy files
+        for filename in datafile_paths:
+            if Path(filename).is_file():
+                if not (fit_dir / Path(filename).name).is_file():
+                    shutil.copyfile(filename, fit_dir / Path(filename).name)
+            else:
+                api_sasview.write_dummy_sans_file(str(fit_dir / Path(filename).name))
 
     # change of cwd is necessary since many fit setup scripts use relative filepaths
     os.chdir(fit_dir)
